Ditch custom type/provider in favor of defined type
This commit (re)ditches the custom type/provider route in favor of the
more manageable defined type. It was exceedingly complex to have a
custom type manage the payload file.  It would have been easy if I could
have added a file resource to the catalog directly from the custom
taype, but alas.